docker_files/requirements_kimi.txt
flask
transformers
torch
accelerate
python-dotenv
Pillow
